Add the box build information to this project
Rather than create a separate git repo (my initial plan), I've placed
all of the stuff for building the neu_libraries/plattr box that this
project now depends on in a subdirectory called box/.  This should make
it easier to work with.  This commit also moves borisrc_base into the
box/ subdirectory because we build boris as part of that step rather
than provisioning, and we also update the .gitignore file to not include
any .box files generated by vagrant package.