Add files necessary to get passenger working
Two things:

1. The way that oh-my-zsh copies the entire PATH variable into
your .zshrc file actually breaks RVM, which is a known issue that looks
like it might never be fixed.  To fix this, I'm adding a zshrc file to
the base box build that comments out the offending PATH dump and more
conservatively just adds the two nonstandard path directories that we
need.

2. While passenger comes with the ability to run its configuration
script silently with the --auto flag, this doesn't handle updating your
httpd.conf in any way, so we copy in an httpd.conf file that has both
the basic Passenger directives we need and the VirtualHost directives
that get tapas/tapas_rails working.